+++ application/modules/store/controllers/admin_items.php
@@ -1,9 +1,9 @@
 <?php
 
+// todo: NO PERMISSIONS!
+
 class Admin_items extends MX_Controller
 {
-	private $sideboxModules;
-
 	public function __construct()
 	{
 		// Make sure to load the administrator library!
@@ -11,6 +11,8 @@ class Admin_items extends MX_Controller
 		$this->load->model('items_model');
 
 		parent::__construct();
+
+		requirePermission("canViewItems");
 	}
 
 	public function index()
@@ -36,8 +38,14 @@ class Admin_items extends MX_Controller
 		$this->administrator->view($content, false, "modules/store/js/admin_items.js");
 	}
 
+	/**
+	 * Create a group that will group some items.
+	 */
 	public function createGroup()
 	{
+		// Check for the permission
+		requirePermission("canAddGroups");
+
 		$data["title"] = $this->input->post("title");
 
 		if(!$data['title'])
@@ -49,14 +57,29 @@ class Admin_items extends MX_Controller
 
 		$this->cache->delete('store_items.cache');
 
+		// Add log
+		$this->logger->createLog('Added item group', $data["title"]);
+
+		$this->plugins->onCreateGroup($data['title']);
+
 		die('window.location.reload(true)');
 	}
 
+	/**
+	 * Add item
+	 */
 	public function create()
 	{
+		// Check for the permission
+		requirePermission("canAddItems");
+
 		if($this->input->post("query"))
 		{
 			$data = $this->getQueryData();
+		}
+		else if($this->input->post("command"))
+		{
+			$data = $this->getCommandData();
 		}
 		else
 		{
@@ -67,9 +90,18 @@ class Admin_items extends MX_Controller
 
 		$this->cache->delete('store_items.cache');
 
+		// Add log
+		$this->logger->createLog('Item added', $data['name']);
+
+		$this->plugins->onAddItem($data);
+
 		die('window.location.reload(true)');
 	}
 
+	/**
+	 * Get the query data
+	 * @return mixed
+	 */
 	private function getQueryData()
 	{
 		$data["name"] = $this->input->post("name");
@@ -85,9 +117,40 @@ class Admin_items extends MX_Controller
 		$data["icon"] = $this->input->post("icon");
 		$data["tooltip"] = 0;
 
+		if(!preg_match("/inv_.+/i", $data["icon"]))
+			$data["icon"] = "inv_misc_questionmark";
+
 		return $data;
 	}
 
+	/**
+	 * Get the command data
+	 * @return mixed
+	 */
+	private function getCommandData()
+	{
+		$data["name"] = $this->input->post("name");
+		$data["description"] = $this->input->post("description");
+		$data["quality"] = $this->input->post("quality");
+		$data["command_need_character"] = ($this->input->post("command_need_character") == "true") ? 1 : 0;
+		$data["command"] = $this->input->post("command");
+		$data["realm"] = $this->input->post("realm");
+		$data["group"] = $this->input->post("group");
+		$data["vp_price"] = $this->input->post("vpCost");
+		$data["dp_price"] = $this->input->post("dpCost");
+		$data["icon"] = $this->input->post("icon");
+		$data["tooltip"] = 0;
+
+		if(!preg_match("/inv_.+/i", $data["icon"]))
+			$data["icon"] = "inv_misc_questionmark";
+
+		return $data;
+	}
+
+	/**
+	 * Get the itemdata
+	 * @return mixed
+	 */
 	private function getItemData()
 	{
 		$data["itemid"] = $this->input->post("itemid");
@@ -96,6 +159,7 @@ class Admin_items extends MX_Controller
 		$data["group"] = $this->input->post("group");
 		$data["vp_price"] = $this->input->post("vpCost");
 		$data["dp_price"] = $this->input->post("dpCost");
+		$data["icon"] = $this->input->post("icon");
 
 		if(!is_numeric(preg_replace("/,/", "", $data["itemid"])))
 		{
@@ -107,7 +171,8 @@ class Admin_items extends MX_Controller
 			$data["name"] = $this->input->post("name");
 			$data["tooltip"] = 0;
 			$data["quality"] = 4;
-			$data["icon"] = "inv_misc_questionmark";
+			if(!preg_match("/inv_.+/i", $data["icon"]))
+				$data["icon"] = "inv_misc_questionmark";
 		}
 		else
 		{
@@ -121,14 +186,22 @@ class Admin_items extends MX_Controller
 			$data["name"] = $item_data['name'];
 			$data["tooltip"] = 1;
 			$data["quality"] = $item_data['Quality'];
-			$data["icon"] = file_get_contents($this->template->page_url."icon/get/".$data["realm"]."/".$data["itemid"]);
+			if(!preg_match("/inv_.+/i", $data["icon"]))
+				$data["icon"] = file_get_contents($this->template->page_url."icon/get/".$data["realm"]."/".$data["itemid"]);
 		}
 
 		return $data;
 	}
 
+	/**
+	 * Load the page to edit the item with the given id.
+	 * @param bool $id
+	 */
 	public function edit($id = false)
 	{
+		// Check for the permission
+		requirePermission("canEditItems");
+
 		if(!is_numeric($id) || !$id)
 		{
 			die();
@@ -163,8 +236,15 @@ class Admin_items extends MX_Controller
 		$this->administrator->view($content, false, "modules/store/js/admin_items.js");
 	}
 
+	/**
+	 * Save the edited details for the given item id.
+	 * @param bool $id
+	 */
 	public function save($id = false)
 	{
+		// Check for the permission
+		requirePermission("canEditItems");
+
 		if(!$id || !is_numeric($id))
 		{
 			die();
@@ -173,6 +253,10 @@ class Admin_items extends MX_Controller
 		if($this->input->post("query"))
 		{
 			$data = $this->getQueryData();
+		}
+		else if($this->input->post("command"))
+		{
+			$data = $this->getCommandData();
 		}
 		else
 		{
@@ -183,11 +267,23 @@ class Admin_items extends MX_Controller
 
 		$this->cache->delete('store_items.cache');
 
+		// Add log
+		$this->logger->createLog('Edited item', $data['name']);
+
+		$this->plugins->onEditItem($id, $data);
+
 		die('window.location="'.$this->template->page_url.'store/admin_items"');
 	}
 
+	/**
+	 * Save a group with the given id
+	 * @param bool $id
+	 */
 	public function saveGroup($id = false)
 	{
+		// Check for the permission
+		requirePermission("canEditGroups");
+
 		if(!$id || !is_numeric($id))
 		{
 			die();
@@ -202,11 +298,19 @@ class Admin_items extends MX_Controller
 
 		$this->items_model->editGroup($id, $data);
 
+		// Add log
+		$this->logger->createLog('Edited item group', $id);
+
+		$this->plugins->onEditGroup($id);
+
 		$this->cache->delete('store_items.cache');
 	}
 
 	public function delete($id = false)
 	{
+		// Check for the permission
+		requirePermission("canRemoveItems");
+
 		if(!$id || !is_numeric($id))
 		{
 			die();
@@ -214,17 +318,29 @@ class Admin_items extends MX_Controller
 
 		$this->items_model->delete($id);
 
+		// Add log
+		$this->logger->createLog('Deleted item', $id);
+
+		$this->plugins->onDeleteItem($id);
+
 		$this->cache->delete('store_items');
 	}
 
 	public function deleteGroup($id = false)
 	{
+		requirePermission("canRemoveGroups");
+
 		if(!$id || !is_numeric($id))
 		{
 			die();
 		}
 
 		$this->items_model->deleteGroup($id);
+
+		// Add log
+		$this->logger->createLog('Deleted group', $id);
+
+		$this->plugins->onDeleteGroup($id);
 
 		$this->cache->delete('store_items');
 	}
